Extra the subview from the parent view
