+++ GymBro/Views/HistoryViews/IntroHistoryView.swift
@@ -111,69 +111,4 @@ class RoutineRowViewModel: ObservableObject {
     
 }
 
-/// The list row sub view for IntroHistoryView
-///
-/// swipteAction, delete and edit
-///
-/// navLink, lead to RoutineDetailView for each
-///
-/// ```
-/// IntroHistoryView_ListRow(element: RoutineSummaryModel)
-/// ```
-struct IntroHistoryView_ListRow: View {
-    
-    /// init parameter
-    @State var element: RoutineSummaryModel
-    
-    /// Call the mock ViewModel
-    @ObservedObject var vm: RoutineRowViewModel = RoutineRowViewModel.instance
-    
-    /// Control for the navigationLink
-    @State private var isActive: Bool = false
-    
-    var body: some View {
-        HStack {
-            Text(element.date)
-            Text(element.routineName)
-            
-            Spacer()
-            Text("\(element.totalExercise) E")
-            Text("\(element.totalSet) S")
-            Text(String(format: "%.2f", element.trainingVolume) + " VI")
-            Image(systemName: "hand.tap")
-                .foregroundColor(.gray)
-            
-            Spacer()
-            Image(systemName: "chevron.left")
-                .foregroundColor(.gray)
-        }
-        .background(
-            NavigationLink(
-                destination: RoutineDetailView(),
-                label: { EmptyView() }
-            )
-            .opacity(0)
-        )
-        .swipeActions(edge: .trailing, allowsFullSwipe: false) {
-            Button(
-                action: {
-                    vm.deleteItem(element: element)
-                },
-                label: {
-                    Label("Delete", systemImage: "trash")
-                }
-            )
-            .tint(.red)
 
-            Button(
-                action: {
-                    vm.editItem(element: element)
-                },
-                label: {
-                    Label("Edit", systemImage: "restart")
-                }
-            )
-            .tint(.green)
-        }
-    }
-}
